added call for travel requirements
